Build fixes (round 1 of >= 4).

+++ programs/Xserver/hw/xfree86/drivers/ati/radeon_common.h
@@ -31,14 +31,13 @@
  * Converted to common header format:
  *   Jens Owen <jens@tungstengraphics.com>
  *
- * $XFree86: xc/programs/Xserver/hw/xfree86/drivers/ati/radeon_common.h,v 1.9 2003/12/13 22:07:22 tsi Exp $
+ * $XFree86: xc/programs/Xserver/hw/xfree86/drivers/ati/radeon_common.h,v 1.10tsi Exp $
  *
  */
 
 #ifndef _RADEON_COMMON_H_
 #define _RADEON_COMMON_H_
 
-#include <inttypes.h>
 #include "xf86drm.h"
 
 /* WARNING: If you change any of these defines, make sure to change
@@ -463,8 +462,8 @@ typedef struct drm_radeon_irq_wait {
  */
 
 typedef struct drm_radeon_set_param {
-	unsigned int param;
-	int64_t      value;
+	unsigned int       param;
+	unsigned long long value;
 } drmRadeonSetParam;
 
 #define RADEON_SETPARAM_FB_LOCATION     1
